app/requirements.txt
pillow
pandas
numpy
scikit-surprise